requirements.txt
flask==3.0.0
moviepy==1.0.3
Pillow==10.2.0
requests==2.31.0
gTTS==2.5.1
gunicorn==21.2.0
